requirements.txt
Flask==1.1.2
numpy==1.21.4
python-dotenv==0.15.0
requests==2.25.0
